requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
attrs==24.2.0
av==14.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
deprecation==2.1.0
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
gotrue==2.11.4
gpt4all==2.8.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.6.1
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.3
openai==1.67.0
packaging==24.2
pillow==11.1.0
postgrest==0.19.3
propcache==0.3.1
protobuf==6.30.0
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.11.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.3
tiktoken
tokenizers==0.21.0
torch==2.2.2
tqdm==4.66.5
transformers==4.49.0
twilio==9.5.0
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.2.3
watchfiles==0.24.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.19.0
